Move new product list components into ./products/list/
Change the products nav item back to /products. Update product list page to use new components. Still have to clean up old components
